+++ masuk/pasien/delete.php
@@ -1,5 +1,5 @@
 <?php
-  $link = mysqli_connect('localhost', 'root', '', 'rs_kdcw') or die(mysqli_error());
+require '../../koneksi.php';
 
   $id = $_GET["No_RM"];
