+++ masuk/pasien/edit.php
@@ -1,5 +1,5 @@
 <?php
-  $link = mysqli_connect('localhost', 'root', '', 'rs_kdcw') or die(mysqli_error());
+require '../../koneksi.php';
   session_start();
   if($_SESSION['login'] != true){
     header('location:../../login.php');
