+++ masuk/rawat/edit.php
@@ -1,8 +1,8 @@
 <?php
-  $link = mysqli_connect('localhost', 'root', '', 'rs_kdcw') or die(mysqli_error());
+require '../../koneksi.php';
   session_start();
   if($_SESSION['login'] != true){
-    header('location:../../login.php');
+    header('location:../../index.php');
   }
   $id = $_GET["no_reg"];
