+++ masuk/rawat/laporan.php
@@ -1,13 +1,9 @@
 <?php
   error_reporting(0);
-  $host     = 'localhost';
-  $user     = 'root';
-  $password = '';
-  $db       = 'rs_kdcw';
-  $link = mysqli_connect('localhost', 'root', '', 'rs_kdcw');
+require '../../koneksi.php';
   session_start();
   if($_SESSION['login'] != true){
-    header('location:../../login.php');
+    header('location:../../index.php');
   }
 
 ?>
