+++ masuk/rawat/simpan.php
@@ -1,5 +1,5 @@
 <?php
-  $link = mysqli_connect('localhost', 'root', '', 'rs_kdcw') or die(mysqli_error($link));
+require '../../koneksi.php';
 
   // Menyimpan data ke dalam variabel
   if (isset($_POST['kirim'])) {
